Fixed broken line number parsing
Surrounded frame line number parsing in try catch to avoid crashes
whenever the stack trace parser fails to properly detect the frame's
line number (either by leaving leading, trailing or in-the-middle spaces
or leaving some text at the end).

If the line number cannot be read, the default value is -1.